CPR-1153 refactored address event listening to use strategy pattern
