+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/cprdomainevents/listeners/AddressDomainEventListener.kt
@@ -1,0 +1,26 @@
+package uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.listeners
+
+import org.springframework.context.annotation.Profile
+import org.springframework.stereotype.Component
+import org.springframework.transaction.event.TransactionalEventListener
+import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.address.AddressCreated
+import uk.gov.justice.digital.hmpps.personrecord.service.cprdomainevents.events.address.AddressUpdated
+
+@Profile("!preprod & !prod")
+@Component
+class AddressDomainEventListener(addressDomainEventStrategies: List<AddressDomainEventStrategy>) {
+
+  private val strategies = addressDomainEventStrategies.associateBy { it.sourceSystemType }
+
+  @TransactionalEventListener
+  fun onAddressCreated(addressCreated: AddressCreated) {
+    val sourceSystem = addressCreated.addressEntity.person!!.sourceSystem
+    strategies[sourceSystem]?.onCreate(addressCreated)
+  }
+
+  @TransactionalEventListener
+  fun onAddressUpdated(addressUpdated: AddressUpdated) {
+    val sourceSystem = addressUpdated.addressEntity.person!!.sourceSystem
+    strategies[sourceSystem]?.onUpdate(addressUpdated)
+  }
+}
